Move examples and benches
