+++ lindera-tokenizer/Cargo.toml
@@ -31,3 +31,11 @@ serde_json = "1.0.91"
 
 lindera-core = { version = "0.23.0", path = "../lindera-core" }
 lindera-dictionary = { version = "0.23.0", path = "../lindera-dictionary" }
+
+[dev-dependencies]
+criterion = { version = "0.4.0", features = ["html_reports"] }
+once_cell = "1.17.0"
+
+[[bench]]
+name = "bench"
+harness = false
