+++ lindera-tokenizer/examples/ko-dic_userdic_example.rs
@@ -1,10 +1,12 @@
-use lindera::LinderaResult;
+use lindera_core::LinderaResult;
 
 fn main() -> LinderaResult<()> {
     #[cfg(feature = "ko-dic")]
     {
+        use std::path::PathBuf;
+
         use lindera_core::viterbi::Mode;
-        use lindera_dictionary::{DictionaryConfig, DictionaryKind};
+        use lindera_dictionary::{DictionaryConfig, DictionaryKind, UserDictionaryConfig};
         use lindera_tokenizer::tokenizer::{Tokenizer, TokenizerConfig};
 
         let dictionary = DictionaryConfig {
@@ -12,9 +14,16 @@ fn main() -> LinderaResult<()> {
             path: None,
         };
 
+        let user_dictionary = Some(UserDictionaryConfig {
+            kind: Some(DictionaryKind::KoDic),
+            path: PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+                .join("../resources")
+                .join("ko-dic_simple_userdic.csv"),
+        });
+
         let config = TokenizerConfig {
             dictionary,
-            user_dictionary: None,
+            user_dictionary: user_dictionary,
             mode: Mode::Normal,
         };
 
@@ -22,7 +31,7 @@ fn main() -> LinderaResult<()> {
         let tokenizer = Tokenizer::from_config(config).unwrap();
 
         // tokenize the text
-        let tokens = tokenizer.tokenize("한국어의형태해석을실시할수있습니다.")?;
+        let tokens = tokenizer.tokenize("하네다공항한정토트백.")?;
 
         // output the tokens
         for token in tokens {
