+++ lindera/Cargo.toml
@@ -48,11 +48,3 @@ lindera-ipadic-builder = { version = "0.23.0", path = "../lindera-ipadic-builder
 lindera-ko-dic-builder = { version = "0.23.0", path = "../lindera-ko-dic-builder" }
 lindera-tokenizer = { version = "0.24.0", path = "../lindera-tokenizer" }
 lindera-unidic-builder = { version = "0.23.0", path = "../lindera-unidic-builder" }
-
-[dev-dependencies]
-criterion = { version = "0.4.0", features = ["html_reports"] }
-once_cell = "1.17.0"
-
-[[bench]]
-name = "bench"
-harness = false
